Break up room commands
